requirements.txt
gspread
google-auth
google-auth-oauthlib
pymupdf
pytesseract
Pillow
opencv-python
pandas
